First Assignment
The first thing to do is to look at the Java files and figure out what this program does. You should be able to run it under Eclipse by adding the 'data' directory to the start up parameters.

Please figure out what each piece of code dose and document it. This includes: file headers, class documentation, method documentation.

Do not create another repository, add branches to this repository and check them in, then request for a push.